Fix Dialyzer error (#115)
Adding a required parameter would make Dialyzer fail with "The created
fun has no local return".  This is true: Maru creates a fun that
handles the case when the parameter has not been passed in, which is
supposed to throw an error.  Let's make Dialyzer not warn about it by
declaring it as "generated code".

Fixes #100.